pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "comfyui-ffmpega"
version = "1.0.0"
description = "Intelligent FFMPEG agent node for ComfyUI with natural language video editing"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "FFMPEGA Team"}
]
keywords = ["comfyui", "ffmpeg", "video-editing", "ai-agent", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.10"
dependencies = [
    "ffmpeg-python>=0.2.0",
    "imageio>=2.31.0",
    "imageio-ffmpeg>=0.4.8",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
ollama = ["ollama>=0.1.0"]
mcp = ["mcp>=1.0.0"]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.urls]
Homepage = "https://github.com/comfyui-ffmpega/ComfyUI-FFMPEGA"
Repository = "https://github.com/comfyui-ffmpega/ComfyUI-FFMPEGA"

[tool.setuptools.packages.find]
where = ["."]
include = ["nodes*", "core*", "skills*", "mcp*", "prompts*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
